requirements.txt
streamlit==1.28.0
pandas==2.1.1
openpyxl==3.1.2
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.0
webdriver-manager==4.0.1
openai==1.3.0
python-dotenv==1.0.0
google-search-results==2.4.2
lxml==4.9.3
fake-useragent==1.4.0
